feat:Eswin Uboot Board level adaptation
Changelogs:
    1. Adapt to Eswin eic7700 chip
    2. Support Eswin EIC7700 EVB Board, SIFIVE hifive_premier_550 Board
    3. Support HDMI, USB, EMMC, SD,SATA,GPIO,MAILBOX,PINCTRL,I2C,SPI,MIPI
    4. Support write eic7700 otp, eic7700 burn and update firmware
    5. add config: CONFIG_CONSOLE_NORMAL=y, CONFIG_VIDEO_ANSI=y, CONFIG_USB_KEYBOARD=y, CONFIG_USE_PREBOOT=y; PREBOOT add "usb start"
    6. es_burn add download/write progress bar
    7. fit(HDMI):Code compliance check

+++ include/net.h
@@ -523,7 +523,7 @@ extern char	net_boot_file_name[1024];/* Boot File name */
 /* Indicates whether the file name was specified on the command line */
 extern bool	net_boot_file_name_explicit;
 /* The actual transferred size of the bootfile (in bytes) */
-extern u32	net_boot_file_size;
+extern u64	net_boot_file_size;
 /* Boot file size in blocks as reported by the DHCP server */
 extern u32	net_boot_file_expected_size_in_blocks;
